% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics1.R
\name{weight}
\alias{weight}
\alias{distance}
\title{Weight Conversion}
\usage{
distance(num, starting, final)
}
\arguments{
\item{num}{A number or a vector for converting.}

\item{starting_unit}{The starting unit of our number.}

\item{final_unit}{The unit we want to convert to.}
}
\value{
A sentence describing the conversion.
}
\description{
This function converts weight between different units
}
\examples{
weight(c(1,2), "g", "tael")
weight(3, "lbs", "kilos")

}
