% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics1.R
\name{convertmetrics}
\alias{convertmetrics}
\alias{distance}
\alias{weight}
\alias{volume}
\alias{temp}
\title{Metrics Conversion}
\usage{
distance(num, starting_unit, final_unit)

weight(num, starting_unit, final_unit)

volume(num, starting_unit, final_unit)

temp(num, starting_unit, final_unit)
}
\arguments{
\item{num}{A number or a vector for converting.}

\item{starting_unit}{The starting_unit unit of our number.}

\item{final_unit}{The unit we want to convert to.}
}
\value{
The converted number or vector.
}
\description{
This function converts metrics between different units. All NA values from the inputs will be discarded.
}
\examples{
distance(6, "yrd", "m")
weight(c(1, 2), "g", "tael")
weight(c(3, NA, 8), "lbs", "kg")
volume(c(10, 20, 30), "gal", "l")
temp(28, "c", "f")

}
